Update Service Fabric sample
