+++ Samples/ServiceFabric/NuGet.Config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <packageSources>
+    <add key="Orleans-local" value="..\..\Artifacts\Debug\" />
+  </packageSources>
+</configuration>
