+++ Samples/ServiceFabric/Stateless/StatelessCalculatorApp/Directory.Build.props
@@ -1,0 +1,2 @@
+<!-- Note that this file is only necessary in this sample project because of how the root Directory.Build.* files interact with the .sfproj -->
+<Project />
