+++ Samples/ServiceFabric/Stateless/StatelessCalculatorApp/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.VisualStudio.Azure.Fabric.MSBuild" version="1.6.0" />
+  <package id="Microsoft.VisualStudio.Azure.Fabric.MSBuild" version="1.6.3" />
 </packages>
