+++ Samples/ServiceFabric/Stateless/StatelessCalculatorService/Program.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
-using System.Fabric;
 using System.Threading;
-using Grains;
 using Microsoft.ServiceFabric.Services.Runtime;
 
 namespace StatelessCalculatorService
@@ -12,29 +10,17 @@ namespace StatelessCalculatorService
         /// <summary>
         /// This is the entry point of the service host process.
         /// </summary>
-        private static void Main()
+        public static void Main()
         {
             try
             {
-                Console.WriteLine(typeof(CalculatorGrain).Name);
-
                 // The ServiceManifest.XML file defines one or more service type names.
                 // Registering a service maps a service type name to a .NET type.
                 // When Service Fabric creates an instance of this service type,
                 // an instance of the class is created in this host process.
-                ServiceEventSource.Current.Message($"Process {Process.GetCurrentProcess().Id} starting");
-                var runtime = FabricRuntime.Create(
-                    () =>
-                    {
-                        Console.WriteLine("Service Fabric Service is exiting.");
-                    });
-                var nodeContext = FabricRuntime.GetNodeContext();
-                var actCont = FabricRuntime.GetActivationContext();
-                Console.WriteLine(nodeContext.ToString() + actCont.ToString() + runtime.ToString());
                 ServiceRuntime.RegisterServiceAsync(
                     "StatelessCalculatorServiceType",
                     context => new StatelessCalculatorService(context)).Wait();
-                ServiceEventSource.Current.Message($"Process {Process.GetCurrentProcess().Id} registered");
 
                 ServiceEventSource.Current.ServiceTypeRegistered(Process.GetCurrentProcess().Id, typeof(StatelessCalculatorService).Name);
